refactor: rename and move chartUtils file
