+++ src/explore-education-statistics-admin/src/pages/release/edit-release/manage-datablocks/components/ChartDataConfiguration.tsx
@@ -10,7 +10,7 @@ import {
   ChartCapabilities,
   colours,
   symbols,
-} from '@common/modules/find-statistics/components/charts/ChartFunctions';
+} from '@common/modules/find-statistics/components/charts/util/chartUtils';
 import {
   ChartSymbol,
   DataSetConfiguration,
