+++ src/explore-education-statistics-admin/src/pages/release/edit-release/manage-datablocks/components/ChartDataSelector.tsx
@@ -11,7 +11,7 @@ import {
   colours,
   generateKeyFromDataSet,
   symbols,
-} from '@common/modules/find-statistics/components/charts/ChartFunctions';
+} from '@common/modules/find-statistics/components/charts/util/chartUtils';
 import {
   ChartDataSet,
   DataSetConfiguration,
