+++ src/explore-education-statistics-common/src/modules/find-statistics/components/DataBlock.tsx
@@ -4,7 +4,7 @@ import TabsSection from '@common/components/TabsSection';
 import ChartRenderer, {
   ChartRendererProps,
 } from '@common/modules/find-statistics/components/ChartRenderer';
-import { parseMetaData } from '@common/modules/find-statistics/components/charts/ChartFunctions';
+import { parseMetaData } from '@common/modules/find-statistics/components/charts/util/chartUtils';
 import TimePeriodDataTableRenderer, {
   Props as TableRendererProps,
 } from '@common/modules/find-statistics/components/TimePeriodDataTableRenderer';
