+++ src/explore-education-statistics-common/src/modules/find-statistics/components/charts/MapBlock.tsx
@@ -8,7 +8,7 @@ import {
   ChartProps,
   createSortedAndMappedDataForAxis,
   generateKeyFromDataSet,
-} from '@common/modules/find-statistics/components/charts/ChartFunctions';
+} from '@common/modules/find-statistics/components/charts/util/chartUtils';
 
 import {
   DataBlockData,
