+++ src/explore-education-statistics-common/src/modules/find-statistics/components/charts/__tests__/HorizontalBarBlock.test.tsx
@@ -3,7 +3,7 @@ import {
   AxesConfiguration,
   ChartMetaData,
   ChartProps,
-} from '@common/modules/find-statistics/components/charts/ChartFunctions';
+} from '@common/modules/find-statistics/components/charts/util/chartUtils';
 
 import { DataBlockData } from '@common/services/dataBlockService';
 import React from 'react';
