+++ src/explore-education-statistics-common/src/modules/find-statistics/components/charts/__tests__/MapBlock.test.tsx
@@ -1,9 +1,8 @@
 /* eslint-disable */
+import testData from '@common/modules/find-statistics/components/charts/__tests__/__data__/testBlockData';
 import React from 'react';
 
 import { render, wait } from 'react-testing-library';
-
-import testData from '@common/modules/find-statistics/components/charts/__tests__/__data__/testBlockData';
 import MapBlock from '../MapBlock';
 
 describe('MapBlock', () => {
