+++ src/explore-education-statistics-common/src/modules/find-statistics/components/charts/util/chartUtils.ts
